requirements.txt
requests
requests_oauthlib

# For testing
pytest
vcrpy
